requirements.txt
accelerate==1.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
attrs==25.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
datasets==3.3.2
dill==0.3.8
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
importlib-metadata==6.11.0
jieba==0.42.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
libretranslatepy==2.1.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.30.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==0.28.0
packaging==23.2
pandas==1.4.2
Pillow==9.4.0
plotly==5.8.0
portalocker==3.1.1
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.23.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
translate==3.6.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.3.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
